Use pytest-order
