+++ pyproject.toml
@@ -40,7 +40,7 @@ poethepoet = "^0.24.2"
 pytest = "^7.2.0"
 pytest-mock = "^3.12.0"
 pytest-subtests = "^0.11.0"
-pytest-dependency = "^0.5.1"
+pytest-order = "^1.1.0"
 
 [tool.poetry.group.dev.dependencies]
 flake8 = { version = "^6.0.0", extras = ["lint"] }
@@ -123,7 +123,7 @@ exclude = [
     "tests/integration/testdata",
 ]
 ignore = ["E203", "E266", "W503"]
-per-file-ignores = ["tests/integration/test_05_docs.py: E501"]
+per-file-ignores = ["tests/integration/test_command_docs.py: E501"]
 
 [tool.pyright]
 pythonVersion = "3.9"
