+++ tests/integration/test_command_docs.py
@@ -389,7 +389,9 @@ def data(file_paths: FilePaths) -> DocsData:
                             "attributes": {
                                 "PROBLEM": "https://onlinejudge.u-aizu.ac.jp/courses/lesson/2/ITP1/1/ITP1_1_A",
                                 "TLE": "0.1",
-                                "links": ["https://onlinejudge.u-aizu.ac.jp/courses/lesson/2/ITP1/1/ITP1_1_A"],
+                                "links": [
+                                    "https://onlinejudge.u-aizu.ac.jp/courses/lesson/2/ITP1/1/ITP1_1_A"
+                                ],
                             },
                             "dependencies": [
                                 {
@@ -1286,214 +1288,220 @@ def check_common(
     assert not list(targets.keys())
 
 
-@pytest.mark.dependency(depends=["resolve_default", "verify_default"], scope="package")
-@pytest.mark.integration
-@pytest.mark.usefixtures("setup_docs")
-def test_with_config(package_dst: pathlib.Path, data: DocsData, subtests: SubTests):
-    destination = package_dst / inspect.stack()[0].function
-    docs_settings_dir = data.root / "docs_settings"
+@pytest.mark.order(-100)
+class TestCommandDocuments:
+    @pytest.mark.integration
+    @pytest.mark.usefixtures("setup_docs")
+    def test_with_config(
+        self,
+        package_dst: pathlib.Path,
+        data: DocsData,
+        subtests: SubTests,
+    ):
+        destination = package_dst / inspect.stack()[0].function
+        docs_settings_dir = data.root / "docs_settings"
 
-    main(
-        [
-            "--docs",
-            docs_settings_dir.as_posix(),
-            "--destination",
-            destination.as_posix(),
+        main(
+            [
+                "--docs",
+                docs_settings_dir.as_posix(),
+                "--destination",
+                destination.as_posix(),
+            ]
+            + data.default_args
+        )
+
+        check_common(destination, data=data, subtests=subtests)
+
+        config_yml = yaml.safe_load((destination / "_config.yml").read_bytes())
+        assert config_yml == {
+            "action_name": "TESTING_WORKFLOW",
+            "basedir": "integration_test_data/",
+            "branch_name": "TESTING_GIT_REF",
+            "description": "My description",
+            "filename-index": True,
+            "highlightjs-style": "vs2015",
+            "plugins": [
+                "jemoji",
+                "jekyll-redirect-from",
+                "jekyll-remote-theme",
+            ],
+            "mathjax": 2,
+            "sass": {"style": "compressed"},
+            "theme": "jekyll-theme-modernist",
+            "icons": {
+                "LIBRARY_ALL_AC": ":heavy_check_mark:",
+                "LIBRARY_ALL_WA": ":x:",
+                "LIBRARY_NO_TESTS": ":warning:",
+                "LIBRARY_PARTIAL_AC": ":heavy_check_mark:",
+                "LIBRARY_SOME_WA": ":question:",
+                "TEST_ACCEPTED": ":100:",
+                "TEST_WAITING_JUDGE": ":warning:",
+                "TEST_WRONG_ANSWER": ":x:",
+            },
+        }
+
+        assert (destination / "static.md").exists()
+        assert (destination / "static.md").read_text(
+            encoding="utf-8"
+        ) == "# Static page\n\nI'm Static!"
+
+        static_dir = docs_settings_dir / "static"
+        for static_file in filter(lambda p: p.is_file(), static_dir.glob("**/*")):
+            assert filecmp.cmp(
+                static_file, destination / static_file.relative_to(static_dir)
+            )
+
+    @pytest.mark.integration
+    @pytest.mark.usefixtures("setup_docs")
+    def test_without_config(
+        self,
+        package_dst: pathlib.Path,
+        data: DocsData,
+        subtests: SubTests,
+    ):
+        destination = package_dst / inspect.stack()[0].function
+
+        main(
+            [
+                "--docs",
+                "testdata/nothing",
+                "--destination",
+                destination.as_posix(),
+            ]
+            + data.default_args
+        )
+
+        check_common(destination, data=data, subtests=subtests)
+
+        config_yml = yaml.safe_load((destination / "_config.yml").read_bytes())
+        assert config_yml == {
+            "action_name": "TESTING_WORKFLOW",
+            "basedir": "integration_test_data/",
+            "branch_name": "TESTING_GIT_REF",
+            "description": '<small>This documentation is automatically generated by <a href="https://github.com/competitive-verifier/competitive-verifier">competitive-verifier/competitive-verifier</a></small>',
+            "filename-index": False,
+            "highlightjs-style": "default",
+            "plugins": [
+                "jemoji",
+                "jekyll-redirect-from",
+                "jekyll-remote-theme",
+            ],
+            "mathjax": 3,
+            "sass": {"style": "compressed"},
+            "theme": "jekyll-theme-minimal",
+            "icons": {
+                "LIBRARY_ALL_AC": ":heavy_check_mark:",
+                "LIBRARY_ALL_WA": ":x:",
+                "LIBRARY_NO_TESTS": ":warning:",
+                "LIBRARY_PARTIAL_AC": ":heavy_check_mark:",
+                "LIBRARY_SOME_WA": ":question:",
+                "TEST_ACCEPTED": ":heavy_check_mark:",
+                "TEST_WAITING_JUDGE": ":warning:",
+                "TEST_WRONG_ANSWER": ":x:",
+            },
+        }
+
+        assert not (destination / "static.md").exists()
+
+        resource_dir = pathlib.Path("src/competitive_verifier_resources/jekyll")
+        for resource_file in filter(lambda p: p.is_file(), resource_dir.glob("**/*")):
+            assert filecmp.cmp(
+                resource_file, destination / resource_file.relative_to(resource_dir)
+            )
+
+    @pytest.mark.integration
+    @pytest.mark.usefixtures("setup_docs")
+    def test_logging_default(
+        self,
+        package_dst: pathlib.Path,
+        data: DocsData,
+        subtests: SubTests,
+        caplog: pytest.LogCaptureFixture,
+    ):
+        caplog.set_level(logging.WARNING)
+        destination = package_dst / inspect.stack()[0].function
+
+        main(
+            [
+                "--docs",
+                "testdata/nothing",
+                "--destination",
+                destination.as_posix(),
+            ]
+            + data.default_args
+        )
+
+        check_common(destination, data=data, subtests=subtests)
+
+        assert caplog.record_tuples == [
+            (
+                "competitive_verifier.documents.job",
+                logging.WARNING,
+                "the `documentation_of` path of dummy/dummy.md is not target: ./dummy.py",
+            )
         ]
-        + data.default_args
-    )
 
-    check_common(destination, data=data, subtests=subtests)
+    @pytest.mark.integration
+    @pytest.mark.usefixtures("setup_docs")
+    def test_logging_include(
+        self,
+        package_dst: pathlib.Path,
+        data: DocsData,
+        subtests: SubTests,
+        caplog: pytest.LogCaptureFixture,
+    ):
+        caplog.set_level(logging.WARNING)
+        destination = package_dst / inspect.stack()[0].function
 
-    config_yml = yaml.safe_load((destination / "_config.yml").read_bytes())
-    assert config_yml == {
-        "action_name": "TESTING_WORKFLOW",
-        "basedir": "integration_test_data/",
-        "branch_name": "TESTING_GIT_REF",
-        "description": "My description",
-        "filename-index": True,
-        "highlightjs-style": "vs2015",
-        "plugins": [
-            "jemoji",
-            "jekyll-redirect-from",
-            "jekyll-remote-theme",
+        main(
+            [
+                "--include",
+                data.targets,
+                "--docs",
+                "testdata/nothing",
+                "--destination",
+                destination.as_posix(),
+            ]
+            + data.default_args
+        )
+
+        check_common(destination, data=data, subtests=subtests)
+
+        assert not caplog.record_tuples
+
+    @pytest.mark.parametrize(
+        "exclude",
+        [
+            ["dummy/"],
+            ["dummy/dummy.md"],
         ],
-        "mathjax": 2,
-        "sass": {"style": "compressed"},
-        "theme": "jekyll-theme-modernist",
-        "icons": {
-            "LIBRARY_ALL_AC": ":heavy_check_mark:",
-            "LIBRARY_ALL_WA": ":x:",
-            "LIBRARY_NO_TESTS": ":warning:",
-            "LIBRARY_PARTIAL_AC": ":heavy_check_mark:",
-            "LIBRARY_SOME_WA": ":question:",
-            "TEST_ACCEPTED": ":100:",
-            "TEST_WAITING_JUDGE": ":warning:",
-            "TEST_WRONG_ANSWER": ":x:",
-        },
-    }
+    )
+    @pytest.mark.integration
+    @pytest.mark.usefixtures("setup_docs")
+    def test_logging_exclude(
+        self,
+        exclude: list[str],
+        package_dst: pathlib.Path,
+        data: DocsData,
+        subtests: SubTests,
+        caplog: pytest.LogCaptureFixture,
+    ):
+        caplog.set_level(logging.WARNING)
+        destination = package_dst / inspect.stack()[0].function
 
-    assert (destination / "static.md").exists()
-    assert (destination / "static.md").read_text(
-        encoding="utf-8"
-    ) == "# Static page\n\nI'm Static!"
-
-    static_dir = docs_settings_dir / "static"
-    for static_file in filter(lambda p: p.is_file(), static_dir.glob("**/*")):
-        assert filecmp.cmp(
-            static_file, destination / static_file.relative_to(static_dir)
+        main(
+            [
+                "--exclude",
+                *exclude,
+                "--docs",
+                "testdata/nothing",
+                "--destination",
+                destination.as_posix(),
+            ]
+            + data.default_args
         )
 
+        check_common(destination, data=data, subtests=subtests)
 
-@pytest.mark.dependency(depends=["resolve_default", "verify_default"], scope="package")
-@pytest.mark.integration
-@pytest.mark.usefixtures("setup_docs")
-def test_without_config(package_dst: pathlib.Path, data: DocsData, subtests: SubTests):
-    destination = package_dst / inspect.stack()[0].function
-
-    main(
-        [
-            "--docs",
-            "testdata/nothing",
-            "--destination",
-            destination.as_posix(),
-        ]
-        + data.default_args
-    )
-
-    check_common(destination, data=data, subtests=subtests)
-
-    config_yml = yaml.safe_load((destination / "_config.yml").read_bytes())
-    assert config_yml == {
-        "action_name": "TESTING_WORKFLOW",
-        "basedir": "integration_test_data/",
-        "branch_name": "TESTING_GIT_REF",
-        "description": '<small>This documentation is automatically generated by <a href="https://github.com/competitive-verifier/competitive-verifier">competitive-verifier/competitive-verifier</a></small>',
-        "filename-index": False,
-        "highlightjs-style": "default",
-        "plugins": [
-            "jemoji",
-            "jekyll-redirect-from",
-            "jekyll-remote-theme",
-        ],
-        "mathjax": 3,
-        "sass": {"style": "compressed"},
-        "theme": "jekyll-theme-minimal",
-        "icons": {
-            "LIBRARY_ALL_AC": ":heavy_check_mark:",
-            "LIBRARY_ALL_WA": ":x:",
-            "LIBRARY_NO_TESTS": ":warning:",
-            "LIBRARY_PARTIAL_AC": ":heavy_check_mark:",
-            "LIBRARY_SOME_WA": ":question:",
-            "TEST_ACCEPTED": ":heavy_check_mark:",
-            "TEST_WAITING_JUDGE": ":warning:",
-            "TEST_WRONG_ANSWER": ":x:",
-        },
-    }
-
-    assert not (destination / "static.md").exists()
-
-    resource_dir = pathlib.Path("src/competitive_verifier_resources/jekyll")
-    for resource_file in filter(lambda p: p.is_file(), resource_dir.glob("**/*")):
-        assert filecmp.cmp(
-            resource_file, destination / resource_file.relative_to(resource_dir)
-        )
-
-
-@pytest.mark.dependency(depends=["resolve_default", "verify_default"], scope="package")
-@pytest.mark.integration
-@pytest.mark.usefixtures("setup_docs")
-def test_logging_default(
-    package_dst: pathlib.Path,
-    data: DocsData,
-    subtests: SubTests,
-    caplog: pytest.LogCaptureFixture,
-):
-    caplog.set_level(logging.WARNING)
-    destination = package_dst / inspect.stack()[0].function
-
-    main(
-        [
-            "--docs",
-            "testdata/nothing",
-            "--destination",
-            destination.as_posix(),
-        ]
-        + data.default_args
-    )
-
-    check_common(destination, data=data, subtests=subtests)
-
-    assert caplog.record_tuples == [
-        (
-            "competitive_verifier.documents.job",
-            logging.WARNING,
-            "the `documentation_of` path of dummy/dummy.md is not target: ./dummy.py",
-        )
-    ]
-
-
-@pytest.mark.dependency(depends=["resolve_default", "verify_default"], scope="package")
-@pytest.mark.integration
-@pytest.mark.usefixtures("setup_docs")
-def test_logging_include(
-    package_dst: pathlib.Path,
-    data: DocsData,
-    subtests: SubTests,
-    caplog: pytest.LogCaptureFixture,
-):
-    caplog.set_level(logging.WARNING)
-    destination = package_dst / inspect.stack()[0].function
-
-    main(
-        [
-            "--include",
-            data.targets,
-            "--docs",
-            "testdata/nothing",
-            "--destination",
-            destination.as_posix(),
-        ]
-        + data.default_args
-    )
-
-    check_common(destination, data=data, subtests=subtests)
-
-    assert not caplog.record_tuples
-
-
-@pytest.mark.parametrize(
-    "exclude",
-    [
-        ["dummy/"],
-        ["dummy/dummy.md"],
-    ],
-)
-@pytest.mark.dependency(depends=["resolve_default", "verify_default"], scope="package")
-@pytest.mark.integration
-@pytest.mark.usefixtures("setup_docs")
-def test_logging_exclude(
-    exclude: list[str],
-    package_dst: pathlib.Path,
-    data: DocsData,
-    subtests: SubTests,
-    caplog: pytest.LogCaptureFixture,
-):
-    caplog.set_level(logging.WARNING)
-    destination = package_dst / inspect.stack()[0].function
-
-    main(
-        [
-            "--exclude",
-            *exclude,
-            "--docs",
-            "testdata/nothing",
-            "--destination",
-            destination.as_posix(),
-        ]
-        + data.default_args
-    )
-
-    check_common(destination, data=data, subtests=subtests)
-
-    assert not caplog.record_tuples
+        assert not caplog.record_tuples
